completed functionality for 3.1

+++ work/week3/ImageInfoViewer/app/src/main/java/ImageInfo.kt
@@ -9,5 +9,5 @@ data class ImageInfo (
     var date: String,
     var share: Boolean,
     var email: String,
-    var rating: Int
+    var rating: Float
 ) : Parcelable
